fix(web): use short description on LandingPage event cards (#44)
All user-visible EventCard usages now consistently show title, metadata
badges (date + registration pills), and the abbreviated short_description.

+++ apps/web/src/pages/LandingPage.tsx
@@ -70,7 +70,7 @@ export const LandingPage: React.FC = () => {
             {!loading && !error && activeEvents.length > 0 && (
               <Stack gap="sm">
                 {visibleActive.map((event) => (
-                  <EventCard key={event.id} event={event} description="long" now={now} />
+                  <EventCard key={event.id} event={event} description="short" now={now} />
                 ))}
                 <Inline>
                   <Link to="/events">See more events</Link>
